requirements.txt
asgiref==3.8.1
click==8.2.1
dj-database-url==3.0.1
Django==4.2
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
h11==0.16.0
Markdown==3.8
mongoengine==0.29.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pymongo==3.11.4
pytz==2025.2
setuptools==80.9.0
sqlparse==0.5.3
uvicorn==0.35.0
whitenoise==6.9.0
django==4.2
daphne>=4.0
channels>=4.0
